🚧 cont: feed, feeddetail 에러 수정.
