+++ src/pages/community/feed/FeedDetail.jsx
@@ -2,77 +2,71 @@ import styled from 'styled-components';
 import { UserInfo } from '@pages/community/user/UserInfo';
 import { ReplyCreate, Replyer } from '@pages/community/feed/ReplyCreate';
 import ReplyList from '@pages/community/feed/ReplyList';
-import { MoreComment } from '@pages/community/feed/Feed';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
+import { ImageArea } from '@pages/community/feed/FeedList';
+import useCustomAxios from '@hooks/useCustomAxios.mjs';
+import { useParams } from 'react-router-dom';
 
-// const replyer = [
-//     {
-//         id: 1,
-//         userId: 'replyer_01',
-//         profileImg: '#',
-//         comment: '사용자 댓글'
-//     },
-// ]
 
-const sample = [
-  {
-    _id: 1,
-    userId: 'user0202',
-    profileImg: '#',
-    post: 'jjj',
-  },
-];
-
-const ImageArea = styled.div`
-  aspect-ratio: 16/9;
-  margin-bottom: 1rem;
-  background-color: #f0f5ed;
-  border-radius: 8px;
-`;
-
-// const Replyer = styled.div`
-//     display: flex;
-//     align-items: center;
-//     gap: 1.2rem;
-//     margin-block: 1rem;
-//     position: relative;
-
-//     & img:first-child {
-//         flex-shrink: 0;
-//         width: 3rem;
-//         height: 3rem;
-//         box-shadow: inset 0 0 20px #335635;
-//         border-radius: 50%;
-//     }
-
-//     & img:last-child {
-//         cursor: pointer;
-//         flex-shrink: 1;
-//         height: 20px;
-//         width: 22px;
-//        }
-// `
-
-function FeedDetail({ item }) {
+function FeedDetail() {
   const [comments, setComments] = useState([]);
-  const { _id, user, profile, content, product, createdAt } = item;
+  const axios = useCustomAxios();
+  const [ data, setData ] = useState();
+  const { _id } = useParams();
+
+  useEffect(() => {
+    fetchDetail();
+  }, []);
+  
+  const fetchDetail = async () => {
+    try {
+      const res = await axios.get(`/posts/${_id}`)
+      console.log(res.data)
+      setData(res.data)
+    }
+     catch (err) {
+     console.log(err)
+     }
+    }
+
+  const item = data?.item;
+     
+
+  
+  // async function fetchList() {
+  //    try {
+  //       const res = await axios.get(`/posts/${_id}`);
+  //        console.log(res)
+  //       // setData(res.data)
+  //    } catch(err) {
+  //        console.log(err);
+  //    }
+  // }
+
 
   function handleAddComment(newComment) {
     setComments([...comments, newComment]);
   }
 
   return (
+    
     <div>
-      <UserInfo profile={profile} userId={user.name} />
-      <div>{post}</div>
-      <ImageArea />
+      {item && (
+        <>
+          <UserInfo profileImg={item.user._id} userId={item.user.name} />
+          <div>{item.content}</div>
+          { item.image || <ImageArea />}
+          
 
-      <Replyer>
-        <ReplyList comments={comments} />
-        <ReplyCreate onAddComment={handleAddComment} item={item} />
-      </Replyer>
-    </div>
+          <Replyer>
+            <ReplyList comments={comments} />
+            <ReplyCreate onAddComment={handleAddComment} item={item} />
+          </Replyer>
+        </>
+      )}
+      </div>
   );
-}
+  }
+
 
 export { FeedDetail, ImageArea };
