+++ src/pages/community/feed/FeedList.jsx
@@ -1,97 +1,134 @@
-import { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import FeedCreate from '@pages/community/feed/FeedCreate';
-import { Feed } from '@pages/community/feed/Feed';
+import styled from "styled-components";
+import { UserInfo } from "@pages/community/user/UserInfo";
+import { Link, useNavigate } from "react-router-dom";
+import iconbookmark from '@assets/images/icon-bookmark.svg'
+import iconbookmarkactive from '@assets/images/icon-bookmark-active.svg'
+import iconsend from '@assets/images/icon-send.svg'
+import { useState } from "react";
+import ReplyList from "@pages/community/feed/ReplyList";
+import { ReplyCreate, Replyer } from "@pages/community/feed/ReplyCreate";
+import FeedDropDown from "@pages/community/feed/FeedDropdown";
+import { FeedDetail } from "@pages/community/feed/FeedDetail";
 
-import useCustomAxios from '@hooks/useCustomAxios.mjs';
-import { useParams } from 'react-router-dom';
+const replyer = [
+    {
+        id: 'i`m user2',
+        userId: 'replyer_01',
+        profileImgUrl: '#',
+    },
+]
 
-const feedList = [
-  {
-    id: 1,
-    userId: 'user_01',
-    profileImgUrl: '#',
-    post: 'ㅁㄴㅇㄹ',
-  },
-  {
-    id: 2,
-    userId: 'user_02',
-    profileImgUrl: '#',
-    post: '안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가',
-  },
-  {
-    id: 3,
-    userId: 'user_03',
-    profileImgUrl: '#',
-    post: 'abcdefddf',
-  },
-  {
-    id: 4,
-    userId: 'user_04',
-    profileImgUrl: '#',
-    post: '안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가',
-  },
-  {
-    id: 5,
-    userId: 'user_05',
-    profileImgUrl: '#',
-    post: 'hit the road',
-  },
-  {
-    id: 6,
-    userId: 'user_06',
-    profileImgUrl: '#',
-    post: 'cocococococococo',
-  },
-  {
-    id: 7,
-    userId: 'user_07',
-    profileImgUrl: '#',
-    post: 'v안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가세요안녕히가',
-  },
-];
+const FeedWrapper = styled.li`
+    max-width: 740px;
+    width: 100%;
+    color: black;
+    font-size: 1.6rem;
+    display: flex;
+    flex-direction: column;
+    margin: 0 auto;
+    position: relative;
 
-const FeedTemplateWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  overflow-y: hidden;
-  padding: 2rem;
+   @media screen and (max-width: 740px) {
+        width: 320px;
+        transition: all 5s easi-in-out;
+   } 
+`
+
+const Post = styled.div`
+    margin-block: 2rem;
+    display: flex;
+    flex-wrap: nowrap;
+    overflow-wrap: break-word;
+`
+
+const StateWrapper = styled.div`
+    display: flex;
+    font-size: 1.8rem;
+    padding: 0.6rem 0;
+    justify-content: space-between;
+
+    & span {
+        color: #727272;
+        font-size: 1.6rem;
+    }
+
+    & img {
+        cursor: pointer;
+        width: 3rem;
+    }
+`
+
+const ImageArea = styled.div`
+  aspect-ratio: 16/9;
+  margin-bottom: 1rem;
+  background-color: #f0f5ed;
+  border-radius: 8px;
 `;
 
-function FeedList() {
-  const [newComment, setNewComment] = useState([]);
-  const axios = useCustomAxios();
-  const [data, setData] = useState();
-  useEffect(() => {
-    fetchList();
-  }, []);
+const MoreComment = styled.div`
+    font-size: 1.6rem;
+    padding: 1rem 0;
+    color: #727272;
+    margin-bottom: 2rem;
+`
 
-  async function fetchList() {
-    try {
-      const res = await axios.get('/posts');
-      console.log(res.data);
-      setData(res.data);
-    } catch (err) {
-      console.log(err);
+function FeedList({ item }){
+    const [ comments, setNewComment ] = useState([]);
+    
+    const [ postId, setPostId ] = useState();
+    const { 
+        _id,
+        name, 
+        profile, 
+        product,
+        content,
+        createdAt } = item;
+    const navigate = useNavigate();
+    
+
+    function handleFeedClick(){
+        console.log(_id, name, profile, content, product, createdAt)
+        navigate(`/community/${_id}`)
     }
-  }
-
-  const item = data?.item;
-  function handleSubmit(e) {
-    e.preventDefault();
-    console.log(newComment);
-    if (newComment.trim() !== '') setNewComment([...newComment, setNewComment]);
-    setNewComment('');
-  }
-
-  return (
-    <FeedTemplateWrapper>
-      {item?.map(item => (
-        <Feed key={item._id} item={item} />
-      ))}
-      <FeedCreate />
-    </FeedTemplateWrapper>
-  );
+    
+    function handleAddComment(newComment){
+        setNewComment([...comments, newComment])
+    }
+    
+    return (
+        <FeedWrapper>
+            <FeedDropDown />
+            <UserInfo profile={profile} userId={name} >
+            </UserInfo >
+            <div onClick={handleFeedClick}>
+                <Post>
+                   <span>{content}</span>
+                </Post>
+                <ImageArea product={product.image}/>
+            </div>
+            <StateWrapper>
+                <span>{createdAt}</span>
+                <img src={iconbookmark} alt="게시글 좋아요 버튼" />
+            </StateWrapper>
+            <UserInfo 
+                profile={profile} 
+                userId={_id} 
+                comment={replyer}/>
+            
+            <ReplyList comments={comments}/>
+                {replyer && replyer.map(t=>
+                <Replyer key={t.id}>
+                    <ReplyCreate onAddComment={handleAddComment}/>
+                    {/* <img src={iconsend} alt="댓글 등록 버튼" /> */}
+                </Replyer>
+                )}
+            <MoreComment>
+                <Link to={`/community/${_id}`}>
+                    댓글 더보기
+                </Link>
+            </MoreComment>
+        </FeedWrapper>
+    )
 }
-export default FeedList;
+
+export { MoreComment, FeedList, FeedWrapper, ImageArea }
