+++ src/routes.jsx
@@ -1,7 +1,9 @@
 import Layout from '@components/layout/layout/Layout';
 import Community from '@pages/community/Community';
 import { FeedDetail } from '@pages/community/feed/FeedDetail';
+
 import FeedEdit from '@pages/community/feed/FeedEdit';
+
 import { FeedNew } from '@pages/community/feed/FeedNew';
 import EditProfile from '@pages/community/mypage/EditProfile';
 import { MyInfo } from '@pages/community/mypage/MyInfo';
@@ -61,7 +63,7 @@ const router = createBrowserRouter([
         element: <Community />,
       },
       {
-        path: 'community/:id',
+        path: 'community/:_id',
         element: <FeedDetail />,
       },
       {
